Replaced "/n" with "\n" as newline in lore.
The standard convention in computing is for newline to be represented by the sequence "\n". I suspect the call to split was uncooperative when the author tried to make the sequence consistent with convention. The split call needs another set of backslashes in it because the string passed is taken to be a RegEx pattern. Because of this, the string is unescaped twice: once when compiling the source code (I think), and again when compiling the RegEx. So "\\\\n" in source becomes "\\n" in bytecode or whatever, and then in the compiled RegEx it is "\n".