Allow "&" in names and lores
With this change, single ampersands ("&") are changed to section symbols ("§"), and double ampersands ("&&") are changed to single ampersands ("&"), allowing people to use "&" in item names and lores. Pretty sure not even Bukkit allows the use of a literal "&" in strings that are to be colored and/or formatted. But we do.